Reorganize add commands and parsers
