+++ src/test/java/seedu/ddd/logic/commands/add/AddContactCommandTest.java
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.Test;
 import javafx.collections.ObservableList;
 import seedu.ddd.commons.core.GuiSettings;
 import seedu.ddd.logic.Messages;
-import seedu.ddd.logic.commands.AddContactCommand;
 import seedu.ddd.logic.commands.CommandResult;
 import seedu.ddd.logic.commands.exceptions.CommandException;
 import seedu.ddd.model.AddressBook;
