+++ src/test/java/seedu/ddd/logic/parser/add/AddClientCommandParserTest.java
@@ -31,8 +31,7 @@ import static seedu.ddd.testutil.contact.TypicalContacts.VALID_CLIENT;
 import org.junit.jupiter.api.Test;
 
 import seedu.ddd.logic.Messages;
-import seedu.ddd.logic.commands.AddContactCommand;
-import seedu.ddd.logic.parser.AddClientCommandParser;
+import seedu.ddd.logic.commands.add.AddContactCommand;
 import seedu.ddd.logic.parser.CommandParserTestUtil;
 import seedu.ddd.model.common.Name;
 import seedu.ddd.model.common.Tag;
