+++ src/test/java/seedu/ddd/testutil/contact/VendorUtil.java
@@ -8,7 +8,7 @@ import static seedu.ddd.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_SERVICE;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_TAG;
 
-import seedu.ddd.logic.commands.AddContactCommand;
+import seedu.ddd.logic.commands.add.AddContactCommand;
 import seedu.ddd.model.contact.vendor.Vendor;
 
 /**
